+++ frontend/compose.yaml
@@ -4,6 +4,6 @@ services:
       context: .
       dockerfile: Dockerfile
 
-    image: ghcr.io/minbzk/ai-act-beslisboom:main
+    image: ghcr.io/minbzk/ai-act-beslishulp:main
     ports:
       - '9090:80'
