+++ frontend/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="node_modules/@nl-rvo/assets/images/favicon/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Beslisboom</title>
+    <title>Beslishulp</title>
   </head>
 
   <body>
