+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ai-act-decision-tree",
+  "name": "ai-act-beslishulp",
   "version": "0.0.5",
   "private": false,
   "type": "module",
