+++ frontend/script/build.sh
@@ -2,4 +2,4 @@
 
 python script/inject_definitions_in_decision_tree.py
 
-docker build . -t ghcr.io/minbzk/ai-act-beslisboom:main --platform linux/amd64
+docker build . -t ghcr.io/minbzk/ai-act-beslishulp:main --platform linux/amd64
