+++ frontend/script/push.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-docker push ghcr.io/minbzk/ai-act-beslisboom:main
+docker push ghcr.io/minbzk/ai-act-beslishulp:main
