+++ frontend/src/App.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import BeslisboomForm from '@/components/BeslisboomForm.vue'
+import BeslishulpForm from '@/components/BeslishulpForm.vue'
 import '@nl-rvo/component-library-css/dist/index.css'
 import '@nl-rvo/assets/fonts/index.css'
 import '@nl-rvo/assets/icons/index.css'
@@ -8,5 +8,5 @@ import '@nl-rvo/design-tokens/dist/index.css'
 </script>
 
 <template>
-  <BeslisboomForm class="rvo-theme" />
+  <BeslishulpForm class="rvo-theme" />
 </template>
