+++ frontend/src/components/betaversion-label.vue
@@ -15,7 +15,7 @@
           </summary>
           <div class="rvo-expandable-text__details absolute"
                style="margin-left: -38px; margin-top: 10px;padding: 16px; border-radius: 5px; background-color: var(--rvo-alert-warning-background-color); width: 220px; z-index: 2">
-            Deze beslisboom is in ontwikkeling. Alle versies ontstaan op een open manier.
+            Deze beslishulp is in ontwikkeling. Alle versies ontstaan op een open manier.
             <a class="rvo-link" href="https://github.com/MinBZK/ai-act-decisiontree">
               Iedereen
               mag opmerkingen en suggesties doen.
